@@ -1,138 +1,155 @@
-cmake_minimum_required(VERSION 3.22)
-project(Meta_Vision_SolaisNG)
+cmake_minimum_required(VERSION 3.25)
+
+# ---- Project ----
+
+project(
+    Meta_Vision_SolaisNG
+    VERSION 1.0
+    LANGUAGES CXX
+)
+
+# ---- Include guards ----
+
+if(PROJECT_SOURCE_DIR STREQUAL PROJECT_BINARY_DIR)
+    message(
+        FATAL_ERROR
+            "In-source builds not allowed. Please make a new directory (called a build directory) and run CMake from there."
+    )
+endif()
 
 add_compile_definitions(PROJECT_NAME="${PROJECT_NAME}")
 
-set(CMAKE_CXX_STANDARD 17)
+
+if (MINGW)
+    set(CMAKE_CXX_STANDARD 17)
+else ()
+    set(CMAKE_CXX_STANDARD 20)
+endif()
 
 include(cmake/CPM.cmake)
 
-#CPMAddPackage(
-#        NAME stdexec
-#        GITHUB_REPOSITORY NVIDIA/stdexec
-#        GIT_TAG main # This will always pull the latest code from the `main` branch. You may also use a specific release version or tag
-#)
+# ---- Add dependencies via CPM ----
+
+CPMAddPackage("gh:TheLartians/PackageProject.cmake@1.8.0")
 
 CPMAddPackage(
-        NAME fmt
-        GIT_TAG 9.1.0
-        GITHUB_REPOSITORY fmtlib/fmt
+    NAME fmt
+    GIT_TAG 9.1.0
+    GITHUB_REPOSITORY fmtlib/fmt
 )
 
 CPMAddPackage(
-        NAME spdlog
-        GITHUB_REPOSITORY gabime/spdlog
-        GIT_TAG v1.11.0
-        OPTIONS
-            "SPDLOG_FMT_EXTERNAL ON"
+    NAME spdlog
+    GITHUB_REPOSITORY gabime/spdlog
+    GIT_TAG v1.11.0
+    OPTIONS "SPDLOG_FMT_EXTERNAL ON"
 )
 
 # Find Boost
 find_package(Boost COMPONENTS system filesystem)
-if (Boost_FOUND)
+if(Boost_FOUND)
     string(REPLACE ";" ", " Boost_LIBRARIES_STRING "${Boost_LIBRARIES}")
-    #    message("Boost ${Boost_VERSION} found: ${Boost_LIBRARIES_STRING}")
-else ()
+    # message("Boost ${Boost_VERSION} found: ${Boost_LIBRARIES_STRING}")
+else()
     message(WARNING "Boost not found so some targets may not be available")
-endif ()
+endif()
 
 # Find OpenCV. Jetson Nano has OpenCV 4.1.1 installed. Here minimal version 4 is required.
 find_package(OpenCV 4)
-if (OpenCV_FOUND)
-    #    message("OpenCV ${OpenCV_VERSION} found")
-else ()
+if(OpenCV_FOUND)
+    # message("OpenCV ${OpenCV_VERSION} found")
+else()
     message(WARNING "OpenCV not found so some targets may not be available")
-endif ()
+endif()
 
 # Find Protocol Buffers
 include(FindProtobuf)
 find_package(Protobuf)
-if (Protobuf_FOUND)
-    #    message("Protobuf ${Protobuf_VERSION} found")
+if(Protobuf_FOUND)
+    # message("Protobuf ${Protobuf_VERSION} found")
     include_directories(${PROTOBUF_INCLUDE_DIR})
-else ()
+else()
     message(WARNING "Protobuf not found so some targets may not be available")
-endif ()
+endif()
 
 # Find Qt
-if (Qt5_DIR)
+if(Qt5_DIR)
     set(QT_VERSION 5)
     set(QT_LIBS Core Gui Widgets)
     set(QT_LIBS_QUALIFIED Qt5::Core Qt5::Gui Qt5::Widgets)
 
-    find_package(Qt${QT_VERSION} COMPONENTS ${QT_LIBS} REQUIRED)
+    find_package(
+        Qt${QT_VERSION}
+        COMPONENTS ${QT_LIBS}
+        REQUIRED
+    )
 
-    if (Qt5_FOUND)
-        #        message("Qt${QT_VERSION} found")
+    if(Qt5_FOUND)
+        # message("Qt${QT_VERSION} found")
 
         # Prioritize Qt5 package include directories
 
         include_directories(BEFORE "${Qt5_DIR}/../../../include")
 
-        #        foreach (qt_lib ${QT_LIBS_QUALIFIED})
-        #            get_target_property(qt_lib_includes ${qt_lib} INTERFACE_INCLUDE_DIRECTORIES)
-        #            message("${qt_lib_includes}")
-        #            include_directories(BEFORE "${qt_lib_includes}")
-        #        endforeach ()
-    endif ()
-else ()
-    message("Qt5_DIR is not defined, and therefore Qt is disabled. "
-            "If you want to use Qt, you may need to set it "
-            "(-DQt5_DIR=\"path/to/Qt/lib/cmake\" or -DQt5_DIR=/usr/include/{host}/qt{version}/ on Ubuntu)")
-endif ()
+        # foreach (qt_lib ${QT_LIBS_QUALIFIED}) get_target_property(qt_lib_includes ${qt_lib}
+        # INTERFACE_INCLUDE_DIRECTORIES) message("${qt_lib_includes}") include_directories(BEFORE
+        # "${qt_lib_includes}") endforeach ()
+    endif()
+else()
+    message(
+        "Qt5_DIR is not defined, and therefore Qt is disabled. "
+        "If you want to use Qt, you may need to set it "
+        "(-DQt5_DIR=\"path/to/Qt/lib/cmake\" or -DQt5_DIR=/usr/include/{host}/qt{version}/ on Ubuntu)"
+    )
+endif()
 
 # Find TensorRT and CUDA
 find_package(CUDAToolkit)
-if (CUDAToolkit_FOUND)
+if(CUDAToolkit_FOUND)
     add_compile_definitions(ON_JETSON)
     enable_language(CUDA)
     include_directories(${CUDAToolkit_INCLUDE_DIRS})
     link_directories(${CUDAToolkit_LIBRARY_DIR})
-else ()
+else()
     message(WARNING "CUDA and TensorRT not found so some targets may not be available")
 endif()
 
 # Find GStreamer
 find_package(PkgConfig)
-if (PkgConfig_FOUND)
+if(PkgConfig_FOUND)
     pkg_check_modules(GSTREAMER gstreamer-1.0)
     pkg_check_modules(GSTREAMER_APP gstreamer-app-1.0)
     pkg_check_modules(GSTREAMER_AUDIO gstreamer-audio-1.0)
     pkg_check_modules(GSTREAMER_PBUTILS gstreamer-pbutils-1.0)
     pkg_check_modules(GSTREAMER_FFT gstreamer-fft-1.0)
 
-    if (GSTREAMER_FOUND AND GSTREAMER_APP_FOUND AND GSTREAMER_AUDIO_FOUND AND GSTREAMER_PBUTILS_FOUND AND GSTREAMER_FFT_FOUND)
+    if(GSTREAMER_FOUND
+       AND GSTREAMER_APP_FOUND
+       AND GSTREAMER_AUDIO_FOUND
+       AND GSTREAMER_PBUTILS_FOUND
+       AND GSTREAMER_FFT_FOUND
+    )
         include_directories(
-                ${GSTREAMER_INCLUDE_DIRS}
-                ${GSTREAMER_APP_INCLUDE_DIRS}
-                ${GSTREAMER_AUDIO_INCLUDE_DIRS}
-                ${GSTREAMER_PBUTILS_INCLUDE_DIRS}
-                ${GSTREAMER_FFT_INCLUDE_DIRS}
+            ${GSTREAMER_INCLUDE_DIRS} ${GSTREAMER_APP_INCLUDE_DIRS} ${GSTREAMER_AUDIO_INCLUDE_DIRS}
+            ${GSTREAMER_PBUTILS_INCLUDE_DIRS} ${GSTREAMER_FFT_INCLUDE_DIRS}
         )
         link_directories(
-                ${GSTREAMER_LIBRARY_DIRS}
-                ${GSTREAMER_APP_LIBRARY_DIRS}
-                ${GSTREAMER_AUDIO_LIBRARY_DIRS}
-                ${GSTREAMER_PBUTILS_LIBRARY_DIRS}
-                ${GSTREAMER_FFT_LIBRARY_DIRS}
+            ${GSTREAMER_LIBRARY_DIRS} ${GSTREAMER_APP_LIBRARY_DIRS} ${GSTREAMER_AUDIO_LIBRARY_DIRS}
+            ${GSTREAMER_PBUTILS_LIBRARY_DIRS} ${GSTREAMER_FFT_LIBRARY_DIRS}
         )
-    else ()
+    else()
         unset(GSTREAMER_FOUND)
-    endif ()
-endif ()
+    endif()
+endif()
 
 if(MINGW)
-    find_library(
-            mingw_winsock
-            libws2_32.a
-    )
-    find_library(
-            mingw_mswsock
-            libmswsock.a
-    )
+    find_library(mingw_winsock libws2_32.a)
+    find_library(mingw_mswsock libmswsock.a)
 
-    MESSAGE("winsock_fix:\n ${mingw_winsock} and ${mingw_mswsock}\n")
+    message("winsock_fix:\n ${mingw_winsock} and ${mingw_mswsock}\n")
 endif()
+
+string(TOLOWER ${PROJECT_NAME}/version.h VERSION_HEADER_LOCATION)
 
 # Shared include directory
 include_directories(include)
